Preserve the BCH type across the QECCore move
