+++ lib/QECCore/ext/QECCoreNemoExt/QECCoreNemoExt.jl
@@ -7,12 +7,11 @@ import Nemo
 import Nemo: GF, gen, matrix, rank, transpose, polynomial_ring, evaluate, FqFieldElem,
     FqPolyRingElem, degree, is_irreducible, gcd, derivative, inv, coeff, is_monic, one,
     minpoly, lcm, is_zero, finite_field
-import QECCore: code_k, parity_matrix_x, parity_matrix_z, parity_matrix, generator_polynomial
+import QECCore: AbstractPolynomialCode, BCH, code_k, code_n, generator_polynomial,
+    parity_matrix, parity_matrix_x, parity_matrix_z, random_Goppa_code
 
 import Random
 import Random: MersenneTwister, GLOBAL_RNG, AbstractRNG, rand
-
-import QECCore: random_Goppa_code, code_k, code_n
 
 function QECCore.code_k(c::AbstractCSSCode)
     n = code_n(c)
